automatically backed up by learn

+++ lib/bachelor.rb
@@ -49,14 +49,14 @@ def get_average_age_for_season(data, season)
    people_counter = 0
    
    data[season].each do |contestants_array|
-     total_age += contestants_array["age"].to_f
+     total_age += contestants_array["age"].to_i
      people_counter += 1.0 
    end 
    
    puts total_age
    puts people_counter
 
-  average_age = ((total_age/people_counter)*100).round /100
+  average_age = ((total_age / people_counter)*100).round /100
    
    return average_age
 end
